Implement dummy jogging
Add Jog controls. Current implementation allows to perform the
Jog using Grbl '$J' command.

Jog can be done using Jog arrow control or by entering the Machine
coordinates to the respective fields.

Before Joggging home cycle should be done, since jogging in currrent
implementation only works with the known machine boundaries.

Refactor DigitalNumberDisplay component. Make allow using the different
font size for decimal and integer parts. Also move calculations to
C++, to achieve better precision.

+++ src/grbl/qtgrblcommon.h
@@ -47,6 +47,6 @@ namespace QtGrbl {
     template<typename T>
     QString enumToString(T value, bool isValid = true)
     {
-        return isValid ? QMetaEnum::fromType<T>().valueToKey(value) : "-";
+        return isValid ? QMetaEnum::fromType<T>().valueToKey(quint64(value)) : "-";
     }
 }
